refactor: changed main dir.
