+++ kadai2/en-ken/cli/cli_test.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"path/filepath"
